Adding new class RequestBlock and refactoring code from RequestHandler related to requestBlock XML to new class.
